Refactor buyer transaction components to utilize withFallback for improved loading states
- Flattened buyer suspense wrappers by integrating `withFallback` into `TransactionList` and `Item` components, enhancing loading state management.
- Removed redundant `Data` and `Pending` components to streamline the structure and improve maintainability.
- Updated `HomeMenu` and various link components to leverage `withFallback`, ensuring consistent loading experiences across the application.
- Enhanced user interaction by adding `data-action` attributes for better accessibility and feedback.
