+++ apps/app/src/app/@user/home/HomeMenu/link/FavouritesLink/FavouritesLink.tsx
@@ -1,21 +1,89 @@
-import type { FC } from "react";
-import { Suspense } from "react";
-import { Data } from "./Data";
-import { Pending } from "./Pending";
+import { useLocale } from "@use-pico/client/hook";
+import { ChevronRightIcon } from "@use-pico/client/icon";
+import type { MarkSuspense } from "@use-pico/client/type";
+import { LinkTo } from "@use-pico/client/ui/link-to";
+import { Tx } from "@use-pico/client/ui/tx";
+import { withFallback } from "@use-pico/client/utils";
+import { FavouriteIcon } from "@zbav-se.me/ui/icon";
+import { TypoIcon } from "@zbav-se.me/ui/typo";
+import { uiMenuButton } from "@zbav-se.me/ui/ui";
 
 export namespace FavouritesLink {
-	export interface Props extends Omit<Data.Props, "_suspense"> {
+	export interface Props extends Pick<LinkTo.Props, "ui" | "iconProps">, MarkSuspense.Props {
 		//
 	}
 }
 
-export const FavouritesLink: FC<FavouritesLink.Props> = (props) => {
-	return (
-		<Suspense fallback={<Pending {...props} />}>
-			<Data
-				_suspense={"I know"}
+export const FavouritesLink = withFallback(
+	({ _suspense, ...props }: FavouritesLink.Props) => {
+		const locale = useLocale();
+
+		return (
+			<LinkTo
+				data-action={"open favourites"}
+				{...uiMenuButton({
+					className: [],
+				})}
+				icon={FavouriteIcon}
+				to="/$locale/buyer/favourite/list"
+				params={{
+					locale,
+				}}
+				activeProps={uiMenuButton({
+					ui: {
+						tone: "primary",
+						theme: "light",
+					},
+					className: [],
+				})}
+				{...uiMenuButton({
+					ui: {
+						tone: "neutral",
+						theme: "light",
+					},
+					className: [],
+				})}
 				{...props}
-			/>
-		</Suspense>
-	);
-};
+			>
+				<TypoIcon
+					flip
+					icon={ChevronRightIcon}
+					iconProps={{
+						ui: {
+							opacity: "5",
+						},
+					}}
+				>
+					<Tx label="Favourites (label)" />
+				</TypoIcon>
+			</LinkTo>
+		);
+	},
+	(props: Omit<FavouritesLink.Props, "_suspense">) => {
+		const locale = useLocale();
+
+		return (
+			<LinkTo
+				data-action={"open favourites"}
+				{...uiMenuButton({
+					className: [],
+				})}
+				icon={FavouriteIcon}
+				to="/$locale/buyer/favourite/list"
+				params={{
+					locale,
+				}}
+				activeProps={uiMenuButton({
+					ui: {
+						tone: "primary",
+						theme: "light",
+					},
+					className: [],
+				})}
+				{...props}
+			>
+				<Tx label={"Loading... (label)"} />
+			</LinkTo>
+		);
+	},
+);
