+++ apps/app/src/app/@user/home/HomeMenu/link/ListingsLink/ListingsLink.tsx
@@ -1,21 +1,64 @@
-import type { FC } from "react";
-import { Suspense } from "react";
-import { Data } from "./Data";
-import { Pending } from "./Pending";
+import { useLocale } from "@use-pico/client/hook";
+import { CartIcon, ChevronRightIcon } from "@use-pico/client/icon";
+import type { MarkSuspense } from "@use-pico/client/type";
+import { LinkTo } from "@use-pico/client/ui/link-to";
+import { Tx } from "@use-pico/client/ui/tx";
+import { withFallback } from "@use-pico/client/utils";
+import { TypoIcon } from "@zbav-se.me/ui/typo";
+import { uiMenuButton } from "@zbav-se.me/ui/ui";
 
 export namespace ListingsLink {
-	export interface Props extends Omit<Data.Props, "_suspense"> {
+	export interface Props extends Pick<LinkTo.Props, "ui" | "iconProps">, MarkSuspense.Props {
 		//
 	}
 }
 
-export const ListingsLink: FC<ListingsLink.Props> = (props) => {
+export const ListingsLink = withFallback(({ _suspense, ...props }: ListingsLink.Props) => {
+	const locale = useLocale();
+
 	return (
-		<Suspense fallback={<Pending {...props} />}>
-			<Data
-				_suspense={"I know"}
-				{...props}
-			/>
-		</Suspense>
+		<LinkTo
+			data-action={"open listings"}
+			{...uiMenuButton({
+				className: [],
+			})}
+			icon={CartIcon}
+			to="/$locale/buyer/feed/default"
+			params={{
+				locale,
+			}}
+			{...props}
+		>
+			<TypoIcon
+				flip
+				icon={ChevronRightIcon}
+				iconProps={{
+					ui: {
+						opacity: "5",
+					},
+				}}
+			>
+				<Tx label="Listings (label)" />
+			</TypoIcon>
+		</LinkTo>
 	);
-};
+}, (props: Omit<ListingsLink.Props, "_suspense">) => {
+	const locale = useLocale();
+
+	return (
+		<LinkTo
+			data-action={"open listings"}
+			{...uiMenuButton({
+				className: [],
+			})}
+			icon={CartIcon}
+			to="/$locale/buyer/feed/default"
+			params={{
+				locale,
+			}}
+			{...props}
+		>
+			<Tx label={"Loading... (label)"} />
+		</LinkTo>
+	);
+});
